+++ source/Processing.IntegrationTests/TestBase.cs
@@ -106,7 +106,7 @@ namespace Processing.IntegrationTests
 
             _container.ConfigureMoveIn(0, 0, TimeOfDay.Create(0, 0, 0));
 
-            _container.Register<IValidator<ChangeCustomerCharacteristicsRequest>, InputValidationSet>(Lifestyle.Scoped);
+            _container.Register<IValidator<ChangeCustomerMasterDataRequest>, InputValidationSet>(Lifestyle.Scoped);
 
             // Integration event publishing
             _serviceBusSenderFactorySpy = new ServiceBusSenderFactorySpy();
